backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
alembic==1.13.3
mutagen==1.47.0
watchfiles==0.24.0
python-multipart==0.0.9
aiofiles==24.1.0
pillow==10.4.0
pydantic-settings==2.5.2
pydantic==2.9.2
apscheduler==3.10.4
spotipy==2.24.0
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.3.0
python-jose[cryptography]==3.3.0
bcrypt==5.0.0
